add user authorization in task

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 @Injectable()
 export class TasksService {
     constructor(
@@ -12,12 +13,18 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ){}
     
-    async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTaskFilterDto,
+        user: User,
+        ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    getTaskById(id: number): Promise<Task> {
-        const found = this.taskRepository.findOne(id);
+    getTaskById(
+        id: number,
+        user: User,
+        ): Promise<Task> {
+        const found = this.taskRepository.findOne({ where: {id, userId: user.id}});
 
         if(!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -26,19 +33,29 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTasksDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTasksDto);
+    async createTask(
+        createTasksDto: CreateTaskDto,
+        user: User,
+        ): Promise<Task> {
+        return this.taskRepository.createTask(createTasksDto, user);
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User
+        ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;    
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User,
+        ): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id});
         
         if(result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
